+++ lang/cs/layouts.php
@@ -2,6 +2,8 @@
 
 return [
 
+    'i18n.index.description' => 'Rozcestník.',
+    'i18n.index.title' => 'Testovací stránka pro překlady',
     'i18n.missing-all.description' => 'Chybí všechny překlady.',
     'i18n.missing-all.title' => 'Testovací stránka pro překlady',
     'i18n.missing-cs.description' => 'Chybí český překlad.',
@@ -10,5 +12,8 @@ return [
     'i18n.missing-en.title' => 'Testovací stránka pro překlady',
     'i18n.complete.description' => 'Kompletní překlady.',
     'i18n.complete.title' => 'Testovací stránka pro překlady',
+
+    'errors.index.description' => 'Rozcestník.',
+    'errors.index.title' => 'Testovací stránka pro chyby.',
 
 ];