+++ resources/views/errors/401.blade.php
@@ -1,0 +1,5 @@
+@extends('ig-common::errors.minimal')
+
+@section('title', __('ig-common::errors.401'))
+@section('code', '401')
+@section('message', __('ig-common::errors.401_message'))