+++ resources/views/errors/402.blade.php
@@ -1,0 +1,5 @@
+@extends('ig-common::errors.minimal')
+
+@section('title', __('ig-common::errors.402'))
+@section('code', '402')
+@section('message', __('ig-common::errors.402_message'))