+++ resources/views/errors/404.blade.php
@@ -1,0 +1,5 @@
+@extends('ig-common::errors.minimal')
+
+@section('title', __('ig-common::errors.404'))
+@section('code', '404')
+@section('message', __('ig-common::errors.404_message'))