+++ resources/views/errors/429.blade.php
@@ -1,0 +1,5 @@
+@extends('ig-common::errors.minimal')
+
+@section('title', __('ig-common::errors.429'))
+@section('code', '429')
+@section('message', __('ig-common::errors.429_message'))