+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add error handler and error pages.
+- Add errors index page.
+- Add i18n index page.
+
 ## [0.6.3] - 2025-03-14
 
 ### Fixed
@@ -177,6 +185,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [0.6.3]: https://https://github.com/internetguru/laravel-common/compare/v0.6.2...v0.6.3
 [0.6.2]: https://https://github.com/internetguru/laravel-common/compare/v0.6.1...v0.6.2
 [0.6.1]: https://https://github.com/internetguru/laravel-common/compare/v0.6.0...v0.6.1